+++ Sources/SuperwallKit/Paywall/Presentation/Get Paywall/InternalGetPaywall.swift
@@ -7,8 +7,16 @@
 
 import Foundation
 import Combine
+import UIKit
 
 extension Superwall {
+  struct PaywallComponents {
+    let viewController: PaywallViewController
+    let presenter: UIViewController?
+    let rulesOutcome: RuleEvaluationOutcome
+    let debugInfo: [String: Any]
+  }
+
   /// Gets a paywall to present, publishing ``PaywallState`` objects that provide updates on the lifecycle of the paywall.
   ///
   /// - Parameters:
@@ -17,46 +25,18 @@ extension Superwall {
   /// - Returns: A ``PaywallViewController`` to present.
   @discardableResult
   func getPaywall(
-    _ request: PresentationRequest
+    _ request: PresentationRequest,
+    _ publisher: PassthroughSubject<PaywallState, Never> = .init()
   ) async throws -> PaywallViewController {
     do {
-      let publisher: PassthroughSubject<PaywallState, Never> = .init()
-      try await waitToPresent(request, paywallStatePublisher: publisher)
+      let paywallComponents = try await getPaywallComponents(request, publisher)
 
-      let debugInfo = logPresentation(
-        request: request,
-        message: "Called Superwall.shared.getPaywall"
-      )
-
-      let rulesOutcome = try await evaluateRules(from: request)
-
-      confirmHoldoutAssignment(from: rulesOutcome)
-
-      let paywallViewController = try await getPaywallViewController(
-        request: request,
-        rulesOutcome: rulesOutcome,
-        debugInfo: debugInfo,
-        paywallStatePublisher: publisher
-      )
-
-      try await checkSubscriptionStatus(
-        request: request,
-        paywall: paywallViewController.paywall,
-        triggerResult: rulesOutcome.triggerResult,
-        paywallStatePublisher: publisher
-      )
-
-      confirmPaywallAssignment(
-        rulesOutcome.confirmableAssignment,
-        isDebuggerLaunched: request.flags.isDebuggerLaunched
-      )
-
-      await paywallViewController.set(
+      await paywallComponents.viewController.set(
         request: request,
         paywallStatePublisher: publisher,
-        unsavedOccurrence: rulesOutcome.unsavedOccurrence
+        unsavedOccurrence: paywallComponents.rulesOutcome.unsavedOccurrence
       )
-      return paywallViewController
+      return paywallComponents.viewController
     } catch {
       let toObjc = request.flags.type.hasObjcDelegate()
       logErrors(from: request, error)