+++ Tests/SuperwallKitTests/Paywall/Presentation/Internal Presentation/Operators/WaitForSubsStatusAndConfigTests.swift
@@ -9,7 +9,7 @@ import XCTest
 @testable import SuperwallKit
 import Combine
 
-final class WaitToPresentTests: XCTestCase {
+final class WaitForSubsStatusAndConfigTests: XCTestCase {
   var cancellables: [AnyCancellable] = []
   let dependencyContainer = DependencyContainer()
   var identityManager: IdentityManager {
@@ -20,7 +20,7 @@ final class WaitToPresentTests: XCTestCase {
     identityManager.reset(duringIdentify: false)
   }
 
-  func test_waitToPresent_noIdentity_unknownStatus() {
+  func test_waitForSubsStatusAndConfig_noIdentity_unknownStatus() {
     let expectation1 = expectation(description: "Got identity")
     expectation1.isInverted = true
 
@@ -41,7 +41,7 @@ final class WaitToPresentTests: XCTestCase {
     .store(in: &cancellables)
 
     Task {
-      try await Superwall.shared.waitToPresent(
+      try await Superwall.shared.waitForSubsStatusAndConfig(
         request,
         paywallStatePublisher: statePublisher,
         dependencyContainer: dependencyContainer
@@ -73,7 +73,7 @@ final class WaitToPresentTests: XCTestCase {
 
     Task {
       do {
-        try await Superwall.shared.waitToPresent(
+        try await Superwall.shared.waitForSubsStatusAndConfig(
           request,
           paywallStatePublisher: statePublisher,
           dependencyContainer: dependencyContainer
@@ -86,7 +86,7 @@ final class WaitToPresentTests: XCTestCase {
     wait(for: [expectation1, stateExpectation], timeout: 5.5)
   }
 
-  func test_waitToPresent_activeStatus_noConfigEvenAfterDelay() {
+  func test_waitForSubsStatusAndConfig_activeStatus_noConfigEvenAfterDelay() {
     let expectation1 = expectation(description: "Got identity")
 
     let unknownSubscriptionPublisher = CurrentValueSubject<SubscriptionStatus, Never>(SubscriptionStatus.active)
@@ -107,7 +107,7 @@ final class WaitToPresentTests: XCTestCase {
 
     Task {
       do {
-        try await Superwall.shared.waitToPresent(
+        try await Superwall.shared.waitForSubsStatusAndConfig(
           stub,
           paywallStatePublisher: statePublisher,
           dependencyContainer: dependencyContainer
@@ -121,7 +121,7 @@ final class WaitToPresentTests: XCTestCase {
     wait(for: [expectation1, stateExpectation], timeout: 1.1)
   }
 
-  func test_waitToPresent_activeStatus_noConfig_configFailedState() {
+  func test_waitForSubsStatusAndConfig_activeStatus_noConfig_configFailedState() {
     let expectation1 = expectation(description: "Got identity")
 
     let unknownSubscriptionPublisher = CurrentValueSubject<SubscriptionStatus, Never>(SubscriptionStatus.active)
@@ -145,7 +145,7 @@ final class WaitToPresentTests: XCTestCase {
 
     Task {
       do {
-        try await Superwall.shared.waitToPresent(
+        try await Superwall.shared.waitForSubsStatusAndConfig(
           stub,
           paywallStatePublisher: statePublisher,
           dependencyContainer: dependencyContainer
@@ -159,7 +159,7 @@ final class WaitToPresentTests: XCTestCase {
     wait(for: [expectation1, stateExpectation], timeout: 0.1)
   }
 
-  func test_waitToPresent_activeStatus_noConfig_configRetryingState() {
+  func test_waitForSubsStatusAndConfig_activeStatus_noConfig_configRetryingState() {
     let expectation1 = expectation(description: "Got identity")
 
     let unknownSubscriptionPublisher = CurrentValueSubject<SubscriptionStatus, Never>(SubscriptionStatus.active)
@@ -183,7 +183,7 @@ final class WaitToPresentTests: XCTestCase {
 
     Task {
       do {
-        try await Superwall.shared.waitToPresent(
+        try await Superwall.shared.waitForSubsStatusAndConfig(
           stub,
           paywallStatePublisher: statePublisher,
           dependencyContainer: dependencyContainer
@@ -197,7 +197,7 @@ final class WaitToPresentTests: XCTestCase {
     wait(for: [expectation1, stateExpectation], timeout: 0.1)
   }
 
-  func test_waitToPresent_activeStatus_noConfig_hasConfigAfterDelay() {
+  func test_waitForSubsStatusAndConfig_activeStatus_noConfig_hasConfigAfterDelay() {
     let expectation1 = expectation(description: "Got identity")
 
     let unknownSubscriptionPublisher = CurrentValueSubject<SubscriptionStatus, Never>(SubscriptionStatus.active)
@@ -221,7 +221,7 @@ final class WaitToPresentTests: XCTestCase {
 
     Task {
       do {
-        try await Superwall.shared.waitToPresent(
+        try await Superwall.shared.waitForSubsStatusAndConfig(
           stub,
           paywallStatePublisher: statePublisher,
           dependencyContainer: dependencyContainer
@@ -235,7 +235,7 @@ final class WaitToPresentTests: XCTestCase {
     wait(for: [expectation1], timeout: 1.1)
   }
 
-  func test_waitToPresent_noIdentity_activeStatus_hasConfig() async {
+  func test_waitForSubsStatusAndConfig_noIdentity_activeStatus_hasConfig() async {
     let expectation = expectation(description: "Got identity")
     expectation.isInverted = true
 
@@ -257,7 +257,7 @@ final class WaitToPresentTests: XCTestCase {
     identityManager.identitySubject.send(false)
     
     Task {
-      try await Superwall.shared.waitToPresent(request, dependencyContainer: dependencyContainer)
+      try await Superwall.shared.waitForSubsStatusAndConfig(request, dependencyContainer: dependencyContainer)
       expectation.fulfill()
     }
 
@@ -281,14 +281,14 @@ final class WaitToPresentTests: XCTestCase {
     .setting(\.flags.subscriptionStatus, to: unknownSubscriptionPublisher)
 
     Task {
-      try await Superwall.shared.waitToPresent(request, dependencyContainer: dependencyContainer)
+      try await Superwall.shared.waitForSubsStatusAndConfig(request, dependencyContainer: dependencyContainer)
       expectation.fulfill()
     }
 
     wait(for: [expectation], timeout: 0.1)
   }
 
-  func test_waitToPresent_inactiveStatus_configFailed() {
+  func test_waitForSubsStatusAndConfig_inactiveStatus_configFailed() {
     let expectation1 = expectation(description: "Got identity")
     expectation1.isInverted = true
 
@@ -321,7 +321,7 @@ final class WaitToPresentTests: XCTestCase {
     .setting(\.flags.subscriptionStatus, to: unknownSubscriptionPublisher)
 
     Task {
-      try await Superwall.shared.waitToPresent(
+      try await Superwall.shared.waitForSubsStatusAndConfig(
         request,
         paywallStatePublisher: statePublisher,
         dependencyContainer: dependencyContainer
@@ -332,7 +332,7 @@ final class WaitToPresentTests: XCTestCase {
     wait(for: [expectation1, stateExpectation], timeout: 0.1)
   }
 
-  func test_waitToPresent_inactiveStatus_hasConfig_noIdentity() {
+  func test_waitForSubsStatusAndConfig_inactiveStatus_hasConfig_noIdentity() {
     let expectation1 = expectation(description: "Got identity")
     expectation1.isInverted = true
 
@@ -366,7 +366,7 @@ final class WaitToPresentTests: XCTestCase {
       try? await Task.sleep(nanoseconds: 1_000_000_000)
       identityManager.identitySubject.send(false)
 
-      try await Superwall.shared.waitToPresent(
+      try await Superwall.shared.waitForSubsStatusAndConfig(
         request,
         paywallStatePublisher: statePublisher,
         dependencyContainer: dependencyContainer
@@ -377,7 +377,7 @@ final class WaitToPresentTests: XCTestCase {
     wait(for: [expectation1, stateExpectation], timeout: 0.1)
   }
 
-  func test_waitToPresent_inactiveStatus_hasConfig_hasIdentity() {
+  func test_waitForSubsStatusAndConfig_inactiveStatus_hasConfig_hasIdentity() {
     let expectation1 = expectation(description: "Got identity")
 
     let unknownSubscriptionPublisher = CurrentValueSubject<SubscriptionStatus, Never>(SubscriptionStatus.inactive)
@@ -409,7 +409,7 @@ final class WaitToPresentTests: XCTestCase {
       // Sleep to allow reset to complete, then set identity as true.
       identityManager.identitySubject.send(true)
 
-      try await Superwall.shared.waitToPresent(
+      try await Superwall.shared.waitForSubsStatusAndConfig(
         request,
         paywallStatePublisher: statePublisher,
         dependencyContainer: dependencyContainer